Extract shared color helpers to utils.js, hide picker when disconnected
Move hexToRgb, darkenHex, and DEFAULT_DIE_COLOR into utils.js to
eliminate duplication between App.svelte and RollScreen.svelte. Wrap
the BandScreen die color picker in a connected check so it is not
rendered for disconnected users.
